Moved models to models file
